# Make /users etc. twice as fast by preloading roles