@tom93 2023-12-30
Thanks for benchmarking, I'll tone down my claimed speed improvement :)
Might be to do with the number of users, investigating...