@tom93 2023-12-30
Here are the timings in prod, definitely improved but not as much as I originally expected.

Before:
```
Completed 200 OK in 5223ms (Views: 4383.0ms | ActiveRecord: 831.3ms)
Completed 200 OK in 5255ms (Views: 4437.3ms | ActiveRecord: 805.8ms)
Completed 200 OK in 5162ms (Views: 4340.2ms | ActiveRecord: 812.4ms)
Completed 200 OK in 5192ms (Views: 4348.5ms | ActiveRecord: 834.2ms)
```
After:
```
Completed 200 OK in 4196ms (Views: 4051.3ms | ActiveRecord: 97.6ms)
Completed 200 OK in 3934ms (Views: 3812.8ms | ActiveRecord: 110.1ms)
Completed 200 OK in 3832ms (Views: 3741.0ms | ActiveRecord: 80.5ms)
Completed 200 OK in 4109ms (Views: 4016.2ms | ActiveRecord: 82.1ms)
```